Split StringBuilderExtensions out to its own file
