+++ Editor/UI/StringBuilderExtensions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a36ada21dc20ef145b2993b8e7258667
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
